Register the instance area with an "url" property
Location services should be registered with an "url" property. Do this
for the (temporary) instance area (aka Eclipse workspace). Note that
this means that we can no longer create the corresponding directory on
demand. Introduce a framework property to control the creation of the
temporary instance area, and update the version number to reflect that
incompatible change.